@@ -4,16 +4,16 @@ description: Demonstrates how to use the flutter_uploader plugin.
 publish_to: "none"
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: ^3.5.3
 
 dependencies:
   flutter:
     sdk: flutter
-  file_picker: ^4.3.3
-  path_provider: ^2.0.8
-  image_picker: ^0.8.4+6
-  shared_preferences: ^2.0.5
-  flutter_local_notifications: ^9.2.0
+  file_picker: ^8.1.3
+  path_provider: ^2.1.4
+  image_picker: ^1.1.2
+  shared_preferences: ^2.3.2
+  flutter_local_notifications: ^17.2.4
 
 dev_dependencies:
   flutter_lints: ^1.0.4
