@@ -5,18 +5,19 @@ homepage: https://github.com/fluttercommunity/flutter_uploader
 maintainer: Sebastian Roth (@ened)
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: ^3.5.3
   flutter: ">=1.20.0"
 
 dependencies:
-  equatable: ^2.0.0
+  equatable: ^2.0.3
+  file_picker: ^8.1.3
   flutter:
     sdk: flutter
 
 dev_dependencies:
-  build_runner: ^2.1.7
+  build_runner: ^2.4.13
   mockito: ^5.0.17
-  flutter_lints: ^1.0.4
+  flutter_lints: ^4.0.0
   flutter_test:
     sdk: flutter
 
