@@ -12,11 +12,11 @@ import 'package:shared_preferences/shared_preferences.dart';
 
 class UploadScreen extends StatefulWidget {
   const UploadScreen({
-    Key? key,
+    super.key,
     required this.uploader,
     required this.uploadURL,
     required this.onUploadStarted,
-  }) : super(key: key);
+  });
 
   final FlutterUploader uploader;
   final Uri uploadURL;
@@ -66,7 +66,7 @@ class _UploadScreenState extends State<UploadScreen> {
               children: <Widget>[
                 Text(
                   'Configure test Server Behavior',
-                  style: Theme.of(context).textTheme.subtitle1,
+                  style: Theme.of(context).textTheme.displayMedium,
                 ),
                 DropdownButton<ServerBehavior>(
                   items: ServerBehavior.all.map((e) {
@@ -85,7 +85,7 @@ class _UploadScreenState extends State<UploadScreen> {
                 const Divider(),
                 Text(
                   'multipart/form-data uploads',
-                  style: Theme.of(context).textTheme.subtitle1,
+                  style: Theme.of(context).textTheme.displayMedium,
                 ),
                 Wrap(
                   alignment: WrapAlignment.center,
@@ -108,7 +108,7 @@ class _UploadScreenState extends State<UploadScreen> {
                 const Divider(height: 40),
                 Text(
                   'binary uploads',
-                  style: Theme.of(context).textTheme.subtitle1,
+                  style: Theme.of(context).textTheme.displayMedium,
                 ),
                 const Text('this will upload selected files as binary'),
                 Wrap(
@@ -215,7 +215,7 @@ class _UploadScreenState extends State<UploadScreen> {
     const tag = 'upload';
 
     var url = binary
-        ? widget.uploadURL.replace(path: widget.uploadURL.path + 'Binary')
+        ? widget.uploadURL.replace(path: '${widget.uploadURL.path}Binary')
         : widget.uploadURL;
 
     url = url.replace(queryParameters: {
